Fix a null safety bug, Refactor code and create barrel file for more clean codebase look
