+++ lib/core/strings.dart
@@ -1,0 +1,3 @@
+class Strings {
+  static const homeTitle = 'Outsightful Quotes';
+}
