+++ lib/core/values.dart
@@ -1,0 +1,4 @@
+class Values {
+  static const double defaultPadding = 20;
+  static const throttleDuration = Duration(milliseconds: 100);
+}
