+++ lib/quotes/logic/bloc/quote_state.dart
@@ -1,4 +1,3 @@
-// ignore_for_file: public_member_api_docs, sort_constructors_first
 part of 'quote_bloc.dart';
 
 enum QuoteStatus { initial, success, failure }
@@ -6,12 +5,12 @@ enum QuoteStatus { initial, success, failure }
 class QuoteState extends Equatable {
   const QuoteState({
     this.status = QuoteStatus.initial,
-    required this.result,
+    this.result = const Result(pagination: Pagination(currentPage: 0, nextPage: 1, totalPages: 1), quotes: <Quote>[]),
     this.hasReachedMax = false,
   });
 
   final QuoteStatus status;
-  final Result? result;
+  final Result result;
   final bool hasReachedMax;
   @override
   List<Object?> get props => [status, result, hasReachedMax];
@@ -27,4 +26,8 @@ class QuoteState extends Equatable {
       hasReachedMax: hasReachedMax ?? this.hasReachedMax,
     );
   }
+
+  @override
+  String toString() =>
+      'QuoteState {status: $status, result: ${result.quotes.length}, hasReachedMax: $hasReachedMax';
 }
