+++ lib/quotes/logic/cubit/theme_state.dart
@@ -3,8 +3,12 @@ part of 'theme_cubit.dart';
 enum ThemeType { light, dark }
 
 class ThemeState {
+  const ThemeState({required this.themeMode});
+
   final ThemeMode themeMode;
   bool get isDarkTheme => themeMode == ThemeMode.dark;
 
-  const ThemeState({required this.themeMode});
+  @override
+  String toString() =>
+      'ThemeState { themeMode: $themeMode, isDarkTheme: $isDarkTheme}';
 }
