+++ lib/quotes/models/quote.dart
@@ -1,69 +1,5 @@
 import 'dart:convert';
 
-class Result {
-  const Result({required this.pagination, required this.quotes});
-
-  final Pagination pagination;
-  final List<Quote> quotes;
-
-  Result copyWith({
-    Pagination? pagination,
-    List<Quote>? quotes,
-  }) {
-    return Result(
-      pagination: pagination ?? this.pagination,
-      quotes: quotes ?? this.quotes,
-    );
-  }
-
-  factory Result.empty() {
-    return const Result(
-        pagination: Pagination(currentPage: 1, nextPage: 2, totalPages: 10),
-        quotes: <Quote>[]);
-  }
-
-  factory Result.fromMap(Map<String, dynamic> map) {
-    return Result(
-      pagination: Pagination.fromMap(map['pagination'] as Map<String, dynamic>),
-      quotes: List<Quote>.from(
-        (map['data'] as List<dynamic>).map<Quote>(
-          (x) => Quote.fromMap(x as Map<String, dynamic>),
-        ),
-      ),
-    );
-  }
-
-  factory Result.fromJson(String source) =>
-      Result.fromMap(json.decode(source) as Map<String, dynamic>);
-
-  @override
-  String toString() {
-    return 'Quote(pagination: $pagination, data: ${quotes.length})';
-  }
-}
-
-class Pagination {
-  final int currentPage;
-  final int nextPage;
-  final int totalPages;
-  const Pagination({
-    required this.currentPage,
-    required this.nextPage,
-    required this.totalPages,
-  });
-
-  factory Pagination.fromMap(Map<String, dynamic> map) {
-    return Pagination(
-      currentPage: map['currentPage'] as int,
-      nextPage: map['nextPage'] as int,
-      totalPages: map['totalPages'] as int,
-    );
-  }
-
-  factory Pagination.fromJson(String source) =>
-      Pagination.fromMap(json.decode(source) as Map<String, dynamic>);
-}
-
 class Quote {
   final String id;
   final String quoteText;
