+++ lib/quotes/quotes.dart
@@ -1,0 +1,3 @@
+export 'package:outsightful_quotes/quotes/logic/cubit/theme_cubit.dart';
+export 'package:outsightful_quotes/quotes/logic/utils/quote_observer.dart';
+export 'package:outsightful_quotes/quotes/presentation/screens/quote_screen.dart';
